requirements.txt
Django==4.2.3
django-extensions
django-debug-toolbar
django-import-export
django-storages
ipython
Celery
crispy-bootstrap5
django-crispy-forms
requests
django-filter

#flake
flake8
flake8-builtins
flake8-print

